requirements.txt
Flask>=2.0.1
spotipy>=2.19.0
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.0.2
python-dotenv>=0.19.0
requests>=2.26.0
